% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyr_install.R
\name{pyr_install}
\alias{pyr_install}
\title{Install saveModelOutput.py}
\usage{
pyr_install(path, overwrite = TRUE)
}
\arguments{
\item{path}{File path where script should be placed}

\item{overwrite}{Overwrite files if they exist. If this is FALSE and the file exists
an error will be thrown.}
}
\description{
Add the \code{saveModelOutput.py} to a project.
}
